Fixed ravager/stalker blood "nothing blood" bug
The bug applied to all of the creature materials, the blood is just an
example.

It was described as "nothing blood", because the original creature name
was removed, but the new name was not put at the top of the object, meaning
the materials were defined (name) before the creature had a real name.

So it looks like there was a reason to use `GO_TO_START` after all.

Sadly though, `GO_TO` tokens don't work in the creature variation itself,
so I have to copy and paste the names in all the illager beast variants.